#10 option to remove failed command (#11)
* Add option --filter-failed
Enabled filtering of failed commands when listing the history.
Will filter out all commands that had a return code that is not 0.

* Add --find-status arg
When specified will find all commands with the given return code.

+++ src/message.rs
@@ -93,7 +93,7 @@ impl CommandStart {
 pub struct CommandFinished {
     pub session_id: Uuid,
     pub time_stamp: DateTime<Utc>,
-    pub result: usize,
+    pub result: u16,
 }
 
 impl CommandFinished {
